djangorestblog/authentication/urls.py
from django.urls import path
from rest_framework_simplejwt import views as jwt_views

from . import views

urlpatterns = [
    path("", views.HelloAuthView.as_view(), name="hello_auth"),
    path("signup/", view=views.UserCreateView.as_view(), name="sign_up"),
    path("token/", jwt_views.TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", jwt_views.TokenRefreshView.as_view(), name="token_refresh"),
]
