requirements.txt
setuptools==65.6.0

python-dateutil==2.8.2
python-decouple==3.6

# django dependencies

Django==4.1.3
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

psycopg2-binary==2.9.5

# test dependencies
factory-boy==3.2.1
Faker==15.3.1
pytest==6.2.5
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-pythonpath==0.7.4
